+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Encode base32 without padding.
+- Encode base32hex without padding.
+
 ## [0.1.0] - 2021-03-25
 ### Added
 - Encode base32.